setting: 컴포넌트 메인페이지 폴더 생성, 파일 이동
