+++ src/components/mainPage/banner/HeaderBanner.tsx
