+++ src/components/mainPage/banner/KakaoTalkBanner.tsx
