+++ src/components/mainPage/banner/PopupBanner.tsx
