+++ src/components/mainPage/personalSection/PersonalBrowser.tsx
