+++ src/components/mainPage/personalSection/PersonalSection.tsx
