+++ src/components/mainPage/personalSection/PersonalTitle.tsx
