+++ src/components/mainPage/productCard/ProductCarousel.tsx
@@ -5,11 +5,15 @@ import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import { Link } from 'react-router-dom';
+import { getProductData } from './getProductData';
+import { useEffect } from 'react';
 
 const settings = {
   slidesToShow: 4,
   slidesToScroll: 1,
 };
+
+getProductData();
 
 export const ProductCarousel = () => {
   return (
