+++ src/components/mainPage/productCard/getProductData.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
+import { API_URL } from '../../../constants/API_URL';
 
-const URL = 'http://13.125.151.45:8080/api/package/admin';
+const URL = API_URL.GET.ALL_CATEGORY;
 export const getProductData = async () => {
   try {
     const response = await axios.get(URL).then((res) => res.data);
