+++ src/components/mainPage/productCard/productData.ts
