+++ src/components/mainPage/recommended/GroupCarousel.tsx
