+++ src/components/mainPage/recommended/RecommendedByRegion.tsx
@@ -1,13 +1,11 @@
 import styled from 'styled-components';
-import { LAYOUT } from '../../constants/layout';
+import { LAYOUT } from '../../../constants/layout';
 import { ProductCarousel } from '../productCard/ProductCarousel';
 
-export const RecommendedByTheme = () => {
+export const RecommendedByRegion = () => {
   return (
     <RecommendedListContainer>
-      <RecommendedListTitle>
-        특별한 테마 여행을 떠나볼까요?
-      </RecommendedListTitle>
+      <RecommendedListTitle>어디로 떠나세요?</RecommendedListTitle>
       <ProductCarousel />
     </RecommendedListContainer>
   );
