+++ src/components/mainPage/sideBar/KakaoTalkBanner.tsx
