+++ src/pages/main/Main.tsx
@@ -1,10 +1,10 @@
 import Header from '../../components/common/Header';
-import { MainBanner } from '../../components/banner/MainBanner/MainBanner';
-import { SideBar } from '../../components/banner/sideBar/SideBar';
-import { PersonalSection } from '../../components/personalSection/PersonalSection';
-import { RecommendedByGroup } from '../../components/recommended/RecommendedByGroup';
-import { RecommendedByRegion } from '../../components/recommended/RecommendedByRegion';
-import { RecommendedByTheme } from '../../components/recommended/RecommendedByTheme';
+import { MainBanner } from '../../components/mainPage/banner/MainBanner/MainBanner';
+import { SideBar } from '../../components/mainPage/sideBar/SideBar';
+import { PersonalSection } from '../../components/mainPage/personalSection/PersonalSection';
+import { RecommendedByGroup } from '../../components/mainPage/recommended/RecommendedByGroup';
+import { RecommendedByRegion } from '../../components/mainPage/recommended/RecommendedByRegion';
+import { RecommendedByTheme } from '../../components/mainPage/recommended/RecommendedByTheme';
 import Review from '../../components/review/Review';
 import Footer from '../../components/common/Footer';
