+++ src/utils/data.ts
@@ -1,0 +1,16 @@
+import instance from './instance';
+
+export default {
+  getProductData() {
+    return instance({
+      url: '/category',
+      method: 'get',
+    });
+  },
+  postMainData() {
+    return instance({
+      url: '/main',
+      method: 'post',
+    });
+  },
+};
